Generative_ai_AIxUI/requirements.txt
aioboto3==13.0.1
aiobotocore==2.13.0
aiofiles==24.1.0
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
boto3==1.34.106
botocore==1.34.106
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.7
deeplake==3.9.11
dill==0.3.8
distro==1.9.0
frozenlist==1.4.1
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hub==3.0.1
humbug==0.3.2
idna==3.7
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.5
langchain-community==0.2.5
langchain-core==0.2.9
langchain-groq==0.1.5
langchain-openai==0.1.9
langchain-text-splitters==0.2.1
langsmith==0.1.82
lz4==4.3.3
marshmallow==3.21.3
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.35.3
orjson==3.10.5
packaging==24.1
pathos==0.3.2
pillow==10.2.0
pox==0.3.4
ppft==1.7.6.8
pydantic==2.7.4
pydantic_core==2.18.4
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
tenacity==8.4.2
tiktoken==0.7.0
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
wrapt==1.16.0
yarl==1.9.4
